pass text size as float instead of typecasting in DisplayActivity

+++ app/src/main/java/edu/temple/activities/MainActivity.kt
@@ -27,7 +27,7 @@ class MainActivity : AppCompatActivity() {
 
             val lyricsDisplayTextView = findViewById<TextView>(R.id.lyricsDisplayTextView)
             adapter = TextSizeAdapter(textSizes) {
-                startActivity(Intent(this@MainActivity, DisplayActivity::class.java).putExtra(TEXT_SIZE_KEY, it))
+                startActivity(Intent(this@MainActivity, DisplayActivity::class.java).putExtra(TEXT_SIZE_KEY, it.toFloat()))
             }
 
             layoutManager = LinearLayoutManager(this@MainActivity)
